+++ src/app/static/src/app/hintVersion.ts
@@ -1,1 +1,2 @@
-export const currentHintVersion = "1.39.0";
+export const currentHintVersion = "1.41.0";
+